refactor: refactor by seperate the mouse coordinates section in to mouse coordinates file and clean the captuchino.tsx
